+++ Mailer/app/Forms/MMSTemplateFormFactory.php
@@ -4,18 +4,22 @@ declare(strict_types=1);
 namespace Remp\Mailer\Forms;
 
 use Nette\Application\UI\Form;
+use Nette\Forms\Controls\SubmitButton;
 use Nette\Security\User;
 use Nette\Utils\ArrayHash;
 use Remp\MailerModule\Models\Auth\PermissionManager;
+use Remp\MailerModule\Models\Segment\Crm;
 use Remp\MailerModule\Repositories\BatchesRepository;
 use Remp\MailerModule\Repositories\JobsRepository;
 use Remp\MailerModule\Repositories\LayoutsRepository;
 use Remp\MailerModule\Repositories\ListsRepository;
 use Remp\MailerModule\Repositories\TemplatesRepository;
-use Remp\MailerModule\Models\Segment\Crm;
 
 class MMSTemplateFormFactory
 {
+    private const FORM_ACTION_WITH_JOBS_CREATED = 'generate_emails_jobs_created';
+    private const FORM_ACTION_WITH_JOBS_STARTED = 'generate_emails_jobs';
+
     private $activeUsersSegment;
 
     private $inactiveUsersSegment;
@@ -106,13 +110,13 @@ class MMSTemplateFormFactory
         $form->setDefaults($defaults);
 
         if ($this->permissionManager->isAllowed($this->user, 'batch', 'start')) {
-            $withJobs = $form->addSubmit('generate_emails_jobs');
+            $withJobs = $form->addSubmit(self::FORM_ACTION_WITH_JOBS_STARTED);
             $withJobs->getControlPrototype()
                 ->setName('button')
                 ->setHtml('Generate newsletter batch and start sending');
         }
 
-        $withJobsCreated = $form->addSubmit('generate_emails_jobs_created');
+        $withJobsCreated = $form->addSubmit(self::FORM_ACTION_WITH_JOBS_CREATED);
         $withJobsCreated->getControlPrototype()
             ->setName('button')
             ->setHtml('Generate newsletter batch');
@@ -146,7 +150,9 @@ class MMSTemplateFormFactory
             $this->batchesRepository->addTemplate($batch, $mailTemplate);
 
             $batchStatus = BatchesRepository::STATUS_READY_TO_PROCESS_AND_SEND;
-            if ($form['generate_emails_jobs_created']->isSubmittedBy()) {
+            /** @var SubmitButton $withJobsCreatedSubmit */
+            $withJobsCreatedSubmit = $form[self::FORM_ACTION_WITH_JOBS_CREATED];
+            if ($withJobsCreatedSubmit->isSubmittedBy()) {
                 $batchStatus = BatchesRepository::STATUS_CREATED;
             }