+++ Mailer/extensions/mailer-module/src/Forms/SnippetFormFactory.php
@@ -125,7 +125,7 @@ class SnippetFormFactory implements IFormFactory
     {
         // decide if user wants to save or save and leave
         $buttonSubmitted = self::FORM_ACTION_SAVE;
-        /** @var $buttonSaveClose SubmitButton */
+        /** @var SubmitButton $buttonSaveClose */
         $buttonSaveClose = $form[self::FORM_ACTION_SAVE_CLOSE];
         if ($buttonSaveClose->isSubmittedBy()) {
             $buttonSubmitted = self::FORM_ACTION_SAVE_CLOSE;