+++ Mailer/extensions/mailer-module/src/Forms/SourceTemplateFormFactory.php
@@ -116,7 +116,7 @@ class SourceTemplateFormFactory implements IFormFactory
     {
         // decide if user wants to save or save and leave
         $buttonSubmitted = self::FORM_ACTION_SAVE;
-        /** @var $buttonSaveClose SubmitButton */
+        /** @var SubmitButton $buttonSaveClose */
         $buttonSaveClose = $form[self::FORM_ACTION_SAVE_CLOSE];
         if ($buttonSaveClose->isSubmittedBy()) {
             $buttonSubmitted = self::FORM_ACTION_SAVE_CLOSE;